+++ src/day9.rs
@@ -1,6 +1,6 @@
 use crate::common::*;
 
-pub fn run(part: usize) -> usize {
+pub fn run(part: u8) -> u64 {
   let input_lines = include_str!("../inputs/day9.txt");
   run_part(part, input_lines, vec_of_rows_of_digits, part1, part2)
 }
@@ -37,7 +37,7 @@ impl Matrix {
   }
 }
 
-pub fn part1(map: &Vec<Vec<u8>>) -> usize { 
+pub fn part1(map: &Vec<Vec<u8>>) -> u64 { 
   let map = Matrix::new(map.clone());
   let mut sum = 0;
   for_low_points(&map, &mut |_, _, value| {
@@ -46,7 +46,7 @@ pub fn part1(map: &Vec<Vec<u8>>) -> usize {
   sum
 }
 
-pub fn part2(map: &Vec<Vec<u8>>) -> usize {
+pub fn part2(map: &Vec<Vec<u8>>) -> u64 {
   let map = Matrix::new(map.clone());
   let mut basin_sizes = vec![];
   for_low_points(&map, &mut |x, y, _| {
@@ -60,7 +60,7 @@ pub fn part2(map: &Vec<Vec<u8>>) -> usize {
     .product()
 }
 
-fn find_basin_size(map: &Matrix, x: usize, y: usize) -> usize {
+fn find_basin_size(map: &Matrix, x: usize, y: usize) -> u64 {
   use std::collections::*;
   let mut filled = HashSet::new();
   let mut q = VecDeque::new();
@@ -74,13 +74,13 @@ fn find_basin_size(map: &Matrix, x: usize, y: usize) -> usize {
       q.extend(basin);  
     }
   }
-  filled.len()
+  filled.len() as u64
 }
 
-fn for_low_points(map: &Matrix, f: &mut impl FnMut(usize, usize, usize)) {
+fn for_low_points(map: &Matrix, f: &mut impl FnMut(usize, usize, u64)) {
   for y in 0..map.y_max {
     for x in 0..map.x_max {
-      let value = map.at(x, y) as usize;
+      let value = map.at(x, y) as u64;
       let neighbours = map.neighbours(x, y);
       let min_neighbours = neighbours.iter()
         .map(|&(x, y)| map.at(x, y))
