+++ src/main.rs
@@ -3,8 +3,8 @@ fn main() {
   use advent_of_code_2021::*;
 
   let args: Vec<String> = env::args().collect();
-  let day = args[1].parse::<usize>().unwrap();
-  let part = args[2].parse::<usize>().unwrap();
+  let day = args[1].parse::<u32>().unwrap();
+  let part = args[2].parse::<u8>().unwrap();
   let run = match day {
     1 => day1::run,
     2 => day2::run,
